One more fix on .debug getvalue + one wrong dabase type get + Trying to implement MSG_MOVE_TIME_SKIPPED + Don't send empty packet with .debug opcode
Former-commit-id: dc3d5b6b9fb331e0fc796b4735f7ea2ed2b352fe

+++ src/game/SocialMgr.cpp
@@ -313,7 +313,7 @@ PlayerSocial *SocialMgr::LoadFromDB(QueryResult result, uint32 guid)
         Field *fields  = result->Fetch();
 
         friend_guid = fields[0].GetUInt32();
-        flags = fields[1].GetUInt32();
+        flags = fields[1].GetUInt8();
         note = fields[2].GetString();
 
         social->m_playerSocialMap[friend_guid] = FriendInfo(flags, note);
